Refactor multicast-broadcast part
