--- task_2_1_2/src/main/java/kozlov/kirill/sockets/BroadcastSender.java
@@ -1,4 +1,0 @@
-package kozlov.kirill.sockets;
-
-public class BroadcastSender {
-}
