+++ task_2_1_2/src/main/java/kozlov/kirill/sockets/data/PortMessage.java
@@ -1,0 +1,3 @@
+package kozlov.kirill.sockets.data;
+
+public record PortMessage(int port, String signature) {}
